requirements.txt
streamlit~=1.28.2
requests~=2.28.2
pandas~=1.5.3
openpyxl~=3.1.0
pymarc~=4.2.1
pyyaml~=6.0